refactor: products 패키지 분리
