+++ src/main/java/kitchenpos/domain/MenuProduct.java
@@ -3,6 +3,8 @@ package kitchenpos.domain;
 import javax.persistence.*;
 import java.util.UUID;
 
+import kitchenpos.products.domain.Product;
+
 @Table(name = "menu_product")
 @Entity
 public class MenuProduct {
