+++ src/main/java/kitchenpos/products/domain/JpaProductRepository.java
@@ -1,8 +1,11 @@
-package kitchenpos.domain;
+package kitchenpos.products.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
+
+import kitchenpos.products.domain.Product;
+import kitchenpos.products.domain.ProductRepository;
 
 public interface JpaProductRepository extends ProductRepository, JpaRepository<Product, UUID> {
 }
