+++ src/main/java/kitchenpos/products/domain/ProductRepository.java
@@ -1,8 +1,10 @@
-package kitchenpos.domain;
+package kitchenpos.products.domain;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+
+import kitchenpos.products.domain.Product;
 
 public interface ProductRepository {
     Product save(Product product);
