+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -2,8 +2,9 @@ package kitchenpos.application;
 
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuRepository;
-import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductRepository;
+import kitchenpos.products.application.ProductService;
+import kitchenpos.products.domain.Product;
+import kitchenpos.products.domain.ProductRepository;
 import kitchenpos.infra.PurgomalumClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
